Make NewExplore the default
