+++ client/src/lib/navigation/ExploreStack.tsx
@@ -4,12 +4,10 @@ import {
   NativeStackNavigationOptions,
 } from '@react-navigation/native-stack';
 import {ExploreStackProps} from './constants/routes';
-import Explore from '../../routes/screens/Explore/Explore';
-import NewExplore from '../../routes/screens/Explore/NewExplore';
+import NewExplore from '../../routes/screens/Explore/Explore';
 import Collection from '../../routes/screens/Collection/Collection';
 import ExploreTag from '../../routes/screens/Explore/ExploreTag';
 import ExploreCategory from '../../routes/screens/Explore/ExploreCategory';
-import {ENVIRONMENT} from 'config';
 
 const {Navigator, Screen} = createNativeStackNavigator<ExploreStackProps>();
 
@@ -20,10 +18,7 @@ const screenOptions: NativeStackNavigationOptions = {
 
 const ExploreStack = () => (
   <Navigator screenOptions={screenOptions}>
-    <Screen
-      name="Explore"
-      component={ENVIRONMENT !== 'production' ? NewExplore : Explore}
-    />
+    <Screen name="Explore" component={NewExplore} />
     <Screen name="ExploreCategory" component={ExploreCategory} />
     <Screen name="ExploreTag" component={ExploreTag} />
     <Screen name="Collection" component={Collection} />
